src/requirements-dev.txt
fastapi
uvicorn[standard]
pydantic-settings
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pymongo
requests